Allow very-org-config for PRs from forks
The `pull_request` event runs in the context of the PR branch and therefore doesn't have access to repository secrets because all the scripts could be compromised.

`pull_request_target` on the other hand runs in the context of the target branch (=master). We can still checkout the PR branch and get the `config.yaml` file from there (no security risk, because that file alone can't do any harm).